port to f64 to match Python

+++ src/action_parsers/action_parser.rs
@@ -2,6 +2,6 @@ use crate::gamestates::game_state::GameState;
 
 
 pub trait ActionParser {
-    fn get_action_space(&mut self) -> Vec<f32>;
-    fn parse_actions(&mut self, actions: Vec<Vec<f32>>, state: &GameState) -> Vec<Vec<f32>>;
+    fn get_action_space(&mut self) -> Vec<f64>;
+    fn parse_actions(&mut self, actions: Vec<Vec<f64>>, state: &GameState) -> Vec<Vec<f64>>;
 }
