+++ src/reward_functions/common_rewards/ball_goal_rewards.rs
@@ -29,8 +29,8 @@ impl RewardFn for VelocityBallToGoalReward {
         
     }
 
-    fn get_reward(&mut self, player: &PlayerData, state: &GameState, previous_action: Vec<f32>) -> f32 {
-        let objective: Vec<f32>;
+    fn get_reward(&mut self, player: &PlayerData, state: &GameState, previous_action: Vec<f64>) -> f64 {
+        let objective: Vec<f64>;
         if (player.team_num == BLUE_TEAM && !self.own_goal) || (player.team_num == ORANGE_TEAM && self.own_goal) {
             objective = ORANGE_GOAL_BACK.to_vec();
         } else {
@@ -49,7 +49,7 @@ impl RewardFn for VelocityBallToGoalReward {
         }
     }
 
-    fn get_final_reward(&mut self, player: &PlayerData, state: &GameState, previous_action: Vec<f32>) -> f32 {
+    fn get_final_reward(&mut self, player: &PlayerData, state: &GameState, previous_action: Vec<f64>) -> f64 {
         self.get_reward(player, state, previous_action)
     }
 }
