+++ src/reward_functions/common_rewards/player_ball_rewards.rs
@@ -21,7 +21,7 @@ impl VelocityPlayerToBallReward {
 impl RewardFn for VelocityPlayerToBallReward {
     fn reset(&mut self, initial_state: &GameState) {}
 
-    fn get_reward(&mut self, player: &PlayerData, state: &GameState, previous_action: Vec<f32>) -> f32 {
+    fn get_reward(&mut self, player: &PlayerData, state: &GameState, previous_action: Vec<f64>) -> f64 {
         let vel = &player.car_data.linear_velocity;
 
         let pos_diff = element_sub_vec(&state.ball.position, &player.car_data.position);
@@ -36,7 +36,7 @@ impl RewardFn for VelocityPlayerToBallReward {
         }
     }
 
-    fn get_final_reward(&mut self, player: &PlayerData, state: &GameState, previous_action: Vec<f32>) -> f32 {
+    fn get_final_reward(&mut self, player: &PlayerData, state: &GameState, previous_action: Vec<f64>) -> f64 {
         self.get_reward(player, state, previous_action)
     }
 }
